Broaden embedding_renorm_ repro to half dtypes and rename for clarity
The old file name said "zero_dim", which reads as a 0-D (scalar) tensor,
while the case under test is a 2-D weight whose embedding dim is 0. Rename
it after what the kernel actually got wrong: the row width.

Also run every case over float32/float16/bfloat16 (the kernel dispatches
half and bfloat16 too), compare the F.embedding output in addition to the
renormed weight, and note that a zero-width regression aborts the whole
process through an asynchronous device fault.
